BOT 1.1.1/requirements.txt
discord
autopy
opencv-python
numpy
pillow
pywin32
pyautogui
keyboard
pynput
pywinauto
pygetwindow